Replace yazi with fyler for file management (#46)
Although yazi nvim plugin is good and fast, this required "yazi" to be installed in the users system, I want to keep this minimal and not ask users to install file managers and all. So the closest good file manager I found was fyler. It gives oil like file management system and provides tree view like neo-tree, best of both worlds, so replacing yazi with fyler.